# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Visa Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /statement/agents:
    get:
      operationId: listAgents
      tags: [statement]
      summary: List agents with balances (owner/provider only)
      responses:
        "200":
          description: Agents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgentAccount"
    post:
      operationId: createAgent
      tags: [statement]
      summary: Register an agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentAccount"
  /statement/agents/{id}:
    put:
      operationId: updateAgent
      tags: [statement]
      summary: Update an agent (rename re-tags transactions)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentAccount"
    delete:
      operationId: deleteAgent
      tags: [statement]
      summary: Delete an agent (transactions keep the old name)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /statement/agents/{id}/details:
    get:
      operationId: getAgentDetails
      tags: [statement]
      summary: Agent statement — transactions, payments, balance
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Agent details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentDetails"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /statement/agents/{id}/payments:
    post:
      operationId: createAgentPayment
      tags: [statement]
      summary: Record a payment with an agent (optionally issues a voucher)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentPaymentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentPayment"
  /statement/payments/{id}:
    delete:
      operationId: deleteAgentPayment
      tags: [statement]
      summary: Delete an agent payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /statement/clients/details:
    get:
      operationId: getClientDetails
      tags: [statement]
      summary: Client statement — transactions and vouchers by client name
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Client details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientDetails"
  /statement/ledger:
    get:
      operationId: listLedger
      tags: [statement]
      summary: List general ledger entries
      responses:
        "200":
          description: Ledger entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LedgerEntry"
    post:
      operationId: createLedgerEntry
      tags: [statement]
      summary: Add a ledger entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LedgerEntryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LedgerEntry"
  /statement/ledger/{id}:
    delete:
      operationId: deleteLedgerEntry
      tags: [statement]
      summary: Delete a ledger entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /vouchers:
    get:
      operationId: listVouchers
      tags: [vouchers]
      summary: List vouchers
      parameters:
        - name: kind
          in: query
          schema:
            type: string
        - name: party
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Vouchers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Voucher"
    post:
      operationId: createVoucher
      tags: [vouchers]
      summary: Create a standalone voucher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoucherInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Voucher"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    AgentAccount:
      type: object
      required: [id, name, totalSales, paidFrom, paidTo, balance, txCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        totalSales:
          type: number
        paidFrom:
          type: number
        paidTo:
          type: number
        balance:
          type: number
        txCount:
          type: integer
        createdAt:
          type: string
    AgentInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        phone:
          type: string
        notes:
          type: string
    AgentPayment:
      type: object
      required: [id, agentId, amount, direction, paidAt, createdAt]
      properties:
        id:
          type: integer
        agentId:
          type: integer
        amount:
          type: number
        direction:
          type: string
        paidAt:
          type: string
        notes:
          type: ["string", "null"]
        voucherId:
          type: ["integer", "null"]
        createdAt:
          type: string
    AgentPaymentInput:
      type: object
      required: [amount, direction]
      properties:
        amount:
          type: number
        direction:
          type: string
          enum: [from_agent, to_agent]
        paidAt:
          type: string
        notes:
          type: string
        createVoucher:
          type: boolean
    StatementTx:
      type: object
      required: [id, clientName, type, issueDate, salePrice, createdAt]
      properties:
        id:
          type: integer
        clientName:
          type: string
        type:
          type: string
        issueDate:
          type: string
        salePrice:
          type: number
        receivedFromClient:
          type: ["number", "null"]
        createdAt:
          type: string
    AgentDetails:
      type: object
      required: [agent, transactions, payments, vouchers]
      properties:
        agent:
          $ref: "#/components/schemas/AgentAccount"
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/StatementTx"
        payments:
          type: array
          items:
            $ref: "#/components/schemas/AgentPayment"
        vouchers:
          type: array
          items:
            $ref: "#/components/schemas/Voucher"
    ClientAccount:
      type: object
      required: [clientName, totalSales, totalReceived, balance, txCount]
      properties:
        clientName:
          type: string
        phone:
          type: ["string", "null"]
        totalSales:
          type: number
        totalReceived:
          type: number
        balance:
          type: number
        txCount:
          type: integer
    ClientDetails:
      type: object
      required: [account, transactions, vouchers]
      properties:
        account:
          $ref: "#/components/schemas/ClientAccount"
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/StatementTx"
        vouchers:
          type: array
          items:
            $ref: "#/components/schemas/Voucher"
    LedgerEntry:
      type: object
      required: [id, type, amount, description, entryDate, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
        amount:
          type: number
        description:
          type: string
        entryDate:
          type: string
        createdAt:
          type: string
    LedgerEntryInput:
      type: object
      required: [type, amount, description]
      properties:
        type:
          type: string
          enum: [income, expense]
        amount:
          type: number
        description:
          type: string
        entryDate:
          type: string
...
    Voucher:
      type: object
      required: [id, kind, partyType, partyName, amount, voucherDate, createdAt]
      properties:
        id:
          type: integer
        kind:
          type: string
        partyType:
          type: string
        partyName:
          type: string
        amount:
          type: number
        description:
          type: ["string", "null"]
        voucherDate:
          type: string
        agentPaymentId:
          type: ["integer", "null"]
        createdAt:
          type: string
    VoucherInput:
      type: object
      required: [kind, partyType, partyName, amount]
      properties:
        kind:
          type: string
          enum: [receipt, payment]
        partyType:
          type: string
          enum: [agent, client, other]
        partyName:
          type: string
        amount:
          type: number
        description:
          type: string
        voucherDate:
          type: string
...