# lib/api-zod/src/generated/types/agentPaymentInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Visa Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentPaymentInputDirection } from './agentPaymentInputDirection';

export interface AgentPaymentInput {
  amount: number;
  direction: AgentPaymentInputDirection;
  paidAt?: string;
  notes?: string;
  createVoucher?: boolean;
}