# lib/api-zod/src/generated/types/voucher.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Visa Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface Voucher {
  id: number;
  kind: string;
  partyType: string;
  partyName: string;
  amount: number;
  /** @nullable */
  description?: string | null;
  voucherDate: string;
  /** @nullable */
  agentPaymentId?: number | null;
  createdAt: string;
}